Split hooks into separate file
