+++ plugins/techdocs-addons/package.json
@@ -38,7 +38,10 @@
     "@types/react": "^16.13.1 || ^17.0.0",
     "react": "^16.13.1 || ^17.0.0"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "@testing-library/react-hooks": "^7.0.2",
+    "@backstage/test-utils": "^1.0.0"
+  },
   "files": [
     "dist"
   ]
